requirements.txt
django
pillow
django-money
openai
numpy
python-decouple
django-filter